+++ src/main/java/net/geforcemods/securitycraft/blocks/InventoryScannerBlock.java
@@ -238,7 +238,7 @@ public class InventoryScannerBlock extends DisguisableBlock {
 		if (!(level.getBlockEntity(pos) instanceof InventoryScannerBlockEntity be))
 			return 0;
 
-		return be.isModuleEnabled(ModuleType.REDSTONE) && be.isProvidingPower() ? 15 : 0;
+		return be.isProvidingPower() ? 15 : 0;
 	}
 
 	@Override