+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadBlock.java
@@ -74,15 +74,19 @@ public class KeypadBlock extends DisguisableBlock {
 	}
 
 	public void activate(BlockState state, Level level, BlockPos pos, int signalLength) {
-		level.setBlockAndUpdate(pos, state.setValue(POWERED, true));
+		level.setBlockAndUpdate(pos, state.cycle(POWERED));
 		BlockUtils.updateIndirectNeighbors(level, pos, SCContent.KEYPAD.get());
-		level.scheduleTick(pos, this, signalLength);
+
+		if (signalLength > 0)
+			level.scheduleTick(pos, this, signalLength);
 	}
 
 	@Override
 	public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
-		level.setBlockAndUpdate(pos, state.setValue(POWERED, false));
-		BlockUtils.updateIndirectNeighbors(level, pos, SCContent.KEYPAD.get());
+		if (state.getValue(POWERED)) {
+			level.setBlockAndUpdate(pos, state.setValue(POWERED, false));
+			BlockUtils.updateIndirectNeighbors(level, pos, SCContent.KEYPAD.get());
+		}
 	}
 
 	@Override